feat(ipc): type the proposals channels instead of unknown (#1632)
`proposal:list` / `proposal:detail` were declared `unknown[]` / `unknown`
in the typed IPC contract — the one surface central to the "LLM proposes,
human confirms" trust model had no wire type, and the proposals store
cast the result with `as Proposal[]`.

Lift the renderer's `Proposal` interface into `src/shared/proposals.ts`
(the serializable, renderer-safe subset of the main-process Proposal) and
point the contract, the client `ProposalsApi`, and the store at it. The
main handler's richer `Proposal` remains structurally assignable, so the
contract is enforced end-to-end by tsc; the `as Proposal[]` cast is gone.

Type-only change (no runtime behavior); verified by tsc + the existing
proposals store/panel/arrival suites.
